Fix misreporting of set result on single param modes
This gets around an issue when using modes with a single parameter where
the powerbox sometimes reports the requested value of D and other times
the "actual" value (where the actual value is equal to the value of C).

I *think* this is a firmware issue, as I can't see any communication
issues in other modes.